# ctrlc/public/swagger.yaml
openapi: 3.0.0
info:
  title: My API
  description: API Documentation with Swagger UI
  version: 1.0.0
paths:
...
  /api/unSubscribe:
    post:
      summary: Unsubscribe a user from a strategy by setting the subscription status to inactive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sub_id:
                  type: string
                  description: The subscription ID to be updated.
                  example: "64dbd10f8fa8a933dcfc8db2"
              required:
                - sub_id
      responses:
        '200':
          description: Successfully unsubscribed user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "unsubscribe"
        '400':
          description: Missing subscription ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "missing sub_id"
        '500':
          description: Server error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "server error"
  /api/subscribedinfo:
    post:
      summary: Fetch a subscription by its ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: The subscription ID to fetch.
                  example: "64dbd10f8fa8a933dcfc8db2"
              required:
                - id
      responses:
        '200':
          description: Subscription details successfully fetched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    type: object
                    example: { "_id": "64dbd10f8fa8a933dcfc8db2", "status": "active" }
        '400':
          description: Missing subscription ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing token or user ID"
        '500':
          description: Server error occurred.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An error occurred while fetching strategies."
...
  /api/trade-history:
    post:
      summary: Send MT5 ID to Fastify API and return trade-history
      description: This endpoint accepts a `mt5_id`, sends it to a Fastify-based backend, and returns the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mt5_id:
                  type: string
                  example: "123456"
              required:
                - mt5_id
      responses:
        '200':
          description: Success. Returns the response from the Fastify API.
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  data:
                    result: "some_result"
        '400':
          description: Bad Request. Missing `mt5_id`.
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "Missing mt5_id"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "Something went wrong"
...